+++ src/app/globals.css
@@ -2,24 +2,6 @@
 @tailwind components;
 @tailwind utilities;
 
-:root {
-  --background: #ffffff;
-  /* --foreground: #171717; */
-  --font-inter: "Inter", sans-serif;
-}
-
-/* @media (prefers-color-scheme: dark) {
-  :root {
-    --background: #0a0a0a;
-    --foreground: #ededed;
-  }
-} */
-
-body {
-  /*color: var(--foreground);*/
-  background: var(--background);
-  font-family: var(--font-inter);
-}
 
 @layer utilities {
   /*Reglas reutilizables de tailwind usando el @apply*/
